Ponderada 1/checkpoint_1/requirements.txt
flask
Flask-SQLAlchemy
Flask-JWT-Extended
requests
Flask-WTF